requirements.txt
numpy
pandas
dash
uvicorn
emoji-country-flag
pycountry